JarvisPrime/src.v3/requirements.txt
requests
pydantic
PyYAML
fastapi
uvicorn
pytest
httpx
sqlmodel
jinja2
python-multipart
cryptography
